+++ pkg-new/k0s/mock.go
