+++ pkg-new/k0s/testdata/patch-k0s-config-change-external-address.yaml
