+++ pkg-new/k0s/testdata/patch-k0s-config-change-sans.yaml
