+++ pkg-new/k0s/testdata/patch-k0s-config-k0s-storage.yaml
