+++ pkg-new/k0s/testdata/patch-k0s-config-no-embedded-cluster-config.yaml
